+++ include/cllm_neighbor_ops.h
@@ -87,7 +87,7 @@ int cllm_internal_apply_neighbor_influence(
  */
 int cllm_internal_apply_neighbor_influence_all(
     CLLMModel* model,
-    float influence_strength
+    double influence_strength
 );
 
 /**
